pyproject.toml
[project]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = [
    {name = "Emil Huuki",email = "116680128+Emil-06737@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pygame (>=2.6.1,<3.0.0)",
    "invoke (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.8.0"
pylint = "^3.3.6"
autopep8 = "^2.3.2"
pytest-dotenv = "^0.5.2"

